+++ src/globalVariables.h
@@ -8,7 +8,7 @@
 #define GLOBALVARIABLES_H
 
 extern volatile float Vref;                            //Reference Voltage, 3.3v
-extern volatile uint16_t ADCVoltage;             //Global - extern - ADC voltage variable and idk if this is good
+extern volatile float ADCVoltage;             //Global - extern - ADC voltage variable and idk if this is good
 extern volatile bool ADCConversionFlag;
